Revised the structure to make it clearer.
